requirements.txt
anyio==4.13.0
certifi==2026.2.25
contourpy==1.3.3
cycler==0.12.1
elastic-transport==8.17.1
elasticsearch==8.13.0
fonttools==4.62.1
idna==3.11
joblib==1.5.3
kiwisolver==1.5.0
matplotlib==3.10.8
numpy==2.4.4
packaging==26.1
pandas==3.0.2
pillow==12.2.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
scikit-learn==1.8.0
scipy==1.17.1
seaborn==0.13.2
six==1.17.0
sniffio==1.3.1
threadpoolctl==3.6.0
typing_extensions==4.15.0
urllib3==2.6.3
